+++ spec/default_spec.rb
@@ -1,14 +1,14 @@
 require 'chefspec'
 
-describe 'chocolatey-chrome::default' do
+describe 'selenium-chromedriver::default' do
   
   let(:chef_run) do
     runner = ChefSpec::ChefRunner.new('platform' => 'windows', 'version'=> '12.04')
-    runner.converge('chocolatey-chrome::default')
+    runner.converge('selenium-chromedriver::default')
   end
     
-  it 'should include the chocolatey-chrome recipe by default' do
-    expect(chef_run).to include_recipe 'chocolatey-chrome::default'
+  it 'should include the selenium-chromedriver recipe by default' do
+    expect(chef_run).to include_recipe 'selenium-chromedriver::default'
   end
 
 end